playlists implementation

+++ src/commands/erase.ts
@@ -4,8 +4,10 @@ import { Wrapper } from "../structures";
 import { getVideoId, validateURL } from "../utils";
 
 export const settings = {
-	aliases : ["e"],
-	description : "permanently removes song from guild history, if no `[link]` specified it will remove curently playing song",
+	name : "Erase",
+	invokes : ["erase", "e"],
+	description : "Removes song from guild history, "
+				+ "if no `[link]` is specified it will try to remove currently playing song.",
 	usage : "[link]",
 	category : "tracking",
 	list : true
@@ -18,7 +20,7 @@ export const erase = async (ID: string, wrapper: Wrapper, message: Message, args
 
 	if (await db.checkGuild(ID)) {
 		// if no argument is given, remove the currently playing song
-		if (!args[0]) {
+		if (args[0] == undefined) {
 			if (queue) {
 				const song = queue.front();
 				if (song) toRemove = getVideoId(song.url);
@@ -26,7 +28,7 @@ export const erase = async (ID: string, wrapper: Wrapper, message: Message, args
 		}
 		// if argument is given, remove the song with given link
 		else if (!validateURL(args[0])) {
-			message.channel.send({ embeds: [wrapper.messageMenager.invalidArguments("erase", settings.aliases, settings.usage)] });
+			message.channel.send({ embeds: [wrapper.messageMenager.invalidArguments(settings)] });
 			return;
 		}
 		else {
@@ -39,8 +41,10 @@ export const erase = async (ID: string, wrapper: Wrapper, message: Message, args
 		const playlistId = guild?.default_playlist_id;
 		if (!playlistId) return;
 
+		const song = await db.getSong(toRemove);
+		if (song == null) return;
 		db.removeFromPlaylist(playlistId, toRemove);
-		message.channel.send("removed \`" + toRemove  + "\` from \`" + playlistId + "\`");
+		message.channel.send({ embeds: [wrapper.messageMenager.erased(song)] });
 	}
 	else {
 		message.channel.send({ embeds: [wrapper.messageMenager.trackingRequired()] });
